Se crea modulo inventario y se separa de magteria prima
