+++ app.py
@@ -7,9 +7,10 @@ from routes.cliente_routes import cliente_bp
 from routes.auth_routes import auth_bp
 from routes.personal_routes import personal_bp
 from routes.registro_compras_routes import registro_compras_bp
-from routes.inventario_routes import inventario_bp
+from routes.materia_prima_routes import materia_prima_bp
 from models.models import db
 from routes.proveedor_routes import provedor_bp
+from routes.inventario_routes import inventario_bp
 # from models.proveedor_model import db
 # from models.materia_prima_model import db
 app = Flask(__name__)
@@ -29,8 +30,10 @@ app.register_blueprint(cliente_bp)
 app.register_blueprint(auth_bp)
 app.register_blueprint(personal_bp)
 app.register_blueprint(registro_compras_bp)
-app.register_blueprint(inventario_bp)
+app.register_blueprint(materia_prima_bp)
 app.register_blueprint(provedor_bp)
+app.register_blueprint(inventario_bp)
+
 
 if __name__ == '__main__':
     app = create_app()
