+++ requirements.txt
@@ -16,3 +16,5 @@ SQLAlchemy==2.0.27
 typing_extensions==4.9.0
 Werkzeug==3.0.1
 WTForms==3.1.2
+Flask-Login==0.6.3
+Flask-Bcrypt==1.0.1
