+++ templates/layout.html
@@ -8,6 +8,7 @@
     <title>Flowbite Flask</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='css/output.css') }}">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css" />
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
 </head>
 
 <body>
@@ -87,7 +88,7 @@
     </div>
 
     <!-- Contenido principal -->
-    <div class="flex items-center justify-center min-h-screen">
+    <div class="flex items-center justify-center min-h-screen bg-white">
         {% with messages = get_flashed_messages(with_categories=true) %}
             {% if messages %}
                 <div class="absolute top-5 right-5 z-50 space-y-4">
